kraliki-lab/services/agent-board/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
httpx==0.25.1
pyyaml==6.0.1
python-multipart==0.0.6
